% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecodeLatentRepresentation.R
\name{DecodeLatentRepresentation}
\alias{DecodeLatentRepresentation}
\title{SAVERCAT decodes latent low-dimensional representation}
\usage{
DecodeLatentRepresentation(
  SeuratObject,
  assay.name = "RNA",
  slot.name = "counts",
  save.model = NULL,
  alpha = 0.01,
  epochs = 50,
  batch_size = 64,
  verbose = 2
)
}
\arguments{
\item{SeuratObject}{A Seurat object.}

\item{assay.name}{Name of assay where count data is stored. Default is 'RNA'.}

\item{slot.name}{Name of slot where raw count data is stored. Default is 'counts'.}

\item{save.model}{Folder specified to save the model.}

\item{alpha}{Weight of reconstruction loss using only observed covariates B.
Default is 0.01.}

\item{epochs}{Number of epochs to train the model.
If unspecified (\code{Default=NULL}), epochs will be 50.}

\item{batch_size}{Number of samples per gradient update. Default is 64.}

\item{verbose}{Verbosity mode. 0 is silent, 1 is progress bar, 2 is one line
per epoch. Default is 2.}
}
\value{
A Seurat object with training details stored in
\code{SeuratObject@tools$train}. Trained model is stored in the specified folder.
A list of elements in \code{SeuratObject@tools$train} are:
\item{\code{z}}{Mean of posterior normal distribution, which is
the low-dimensional representation of the cells.}
\item{\code{z.var}}{Log of variance of posterior normal distribution.}
\item{\code{z.samp}}{Random sampling from posterior normal distribution of
the latent characterization.}
\item{\code{B}}{Covariates matrix used to train the model.}
\item{\code{sf}}{Cell specific size factors used.}
\item{\code{bad.genes}}{Index of genes that could not be predicted accurately
by the autoencoder based on cross validation results. Instead, for these
genes, denoised SAVER estimate are mean gene expression level across cells.}
\item{\code{time}}{Time of decoder training.}
\item{\code{train1}}{A list that contains training information and model
details when characterizing low-dimensional representation.}
\item{\code{train2}}{A list that contains training information and model
details when decoding the low-dimensional representation .}
}
\description{
This function trains a decoder which takes the low-dimensional
representation and produces covariate-adjusted and denoised expression data.
}
\details{
This function takes a Seurat object with dimensional reduction object in
\code{SeuratObject@reductions$saverc}.
This function aims to train a decoder model and store it in the specified
folder.
Before running this function, \code{LearnLatentRepresentation} should be run.
}
