% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_saver.R
\name{generate_saver}
\alias{generate_saver}
\title{Calculate SAVER-CVAE corrected counts, SAVER estimates and sampled SAVER
estimates.}
\usage{
generate_saver(decoder, x, z, B, bad.genes, subset.genes = NULL)
}
\arguments{
\item{decoder}{Decoder model.}

\item{x}{Expression count matrix. The rows correspond to genes and colomns
correspond to cells.}

\item{z}{Mean of posterior normal distribution, which is the low-dimensional
representation of the cells.}

\item{B}{Observed covariates matrix used to train the model.
The rows correspond to cells.}

\item{bad.genes}{Index of genes that could not be predicted accurately
by the autoencoder based on cross validation results. Instead, for these
genes, denoised SAVER estimate are mean gene expression level across cells.}

\item{subset.genes}{Index of a subset of genes to be predicted.
Default is NULL, then all genes will be predicted.}
}
\value{
A list with the following components
\item{\code{estimate}}{Denoised SAVER estimates based on empirical Bayes
method. The SAVER estimate and posterior distribution can be used to
visualize gene expression and perform trajectory analyses. The SAVER
estimate acts as the denoised estimate of the true expression}
\item{\code{se}}{Standard error of SAVER posterior distribution.}
\item{\code{samp}}{Sampling from the SAVER posterior distribution.
Sampling from the SAVER posterior allows investigation on gene expression
distribution.}
\item{\code{adj.count}}{SAVER-CVAE corrected counts using cdf matching. This
preserves the technical noise characteristics of scRNA-seq expression count
data after adjusting for the observed covariates. This is useful in
downstream scenarios where hypothesis testing on the original count data
is desired, such as in differential expression.}
\item{\code{time}}{Runtime.}
}
\description{
Given trained encoder and decoder model as input, calculate SAVER-CVAE
corrected counts, SAVER estimates and sampled SAVER estimates.
}
