% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess gene expression count matrix}
\usage{
preprocess_data(
  x,
  B = NULL,
  n_hvg = 3000,
  hvg_genes_only = TRUE,
  perc.exp = 0.01
)
}
\arguments{
\item{x}{Expression count matrix. The rows correspond to genes and colomns
correspond to cells.}

\item{B}{Observed covariates matrix. The rows correspond to cells.
Default is NULL, then scaled log library size will be taken as B.}

\item{n_hvg}{Number of highly variable genes to select. Default is 3000.}

\item{hvg_genes_only}{Whether to subset and normalize expression count matrix
using only highly variable genes. TRUE to use only highly variable genes.
FALSE to use genes with non-zero expression in at least perc.exp*100 \% of
cells. Default is TRUE.}

\item{perc.exp}{Good genes are defined as those with a mean expression of
more than perc.exp. We select highly variable genes from good
genes.}
}
\value{
A list with the following components
\item{\code{x}}{A subset of expression count matrix. The rows correspond to
cells and colomns correspond to highly variable genes or genes with a mean
expression of more than perc.exp.}
\item{\code{x.norm}}{Log normailized and standardized gene expression matrix
(cells by genes).}
\item{\code{sf}}{Cell specific size factors used.}
\item{\code{B}}{Observed covariates matrix or scaled log library size.}
\item{\code{hvg}}{Names of highly variable genes.}
}
\description{
This preprocessing function calculates library size, library size
normalization factor, highly variable genes, normalized gene expression
level. These results can be used in training SAVER-CVAE model.
}
\details{
This function normalized gene expression count matrix by size factor, then
log scale and standardize it. Size factor is library size diveded by mean
library size. It calculates highly variable genes per batch using
\code{Seurat::FindVariableFeatures} function and takes the union.
}
