% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_decoder_sc.R
\name{run_decoder_sc}
\alias{run_decoder_sc}
\title{Train a decoder which generates denoised expression matrix}
\usage{
run_decoder_sc(
  x,
  z.mean,
  z.logvar,
  B = NULL,
  sf = NULL,
  epochs = 50,
  alpha = 0.01,
  return_saver = FALSE,
  verbose = 2,
  batch_size = 64,
  save.model = NULL
)
}
\arguments{
\item{x}{Expression count matrix. The rows correspond to genes and colomns
correspond to cells.}

\item{z.mean}{Mean of posterior normal distribution of the latent variable,
which is the low-dimensional representation of the cells.}

\item{z.logvar}{Log of variance of posterior normal distribution of the
latent variable.}

\item{B}{Observed covariates matrix. The rows correspond to cells.
Default is NULL, then scaled log library size will be taken as B.}

\item{sf}{Vector of cell specific size factors. Default is NULL, then scaled
library size will be used.}

\item{epochs}{Number of epochs to train the model. Default is 50.}

\item{alpha}{Weight of reconstruction loss using only observed covariates B.
Default is 0.01.}

\item{return_saver}{Whether to calculate corrected expression counts and
SAVER estimates. Default is FALSE.}

\item{verbose}{Verbosity mode. 0 is silent, 1 is progress bar, 2 is one line
per epoch. Default is 2.}

\item{batch_size}{Number of samples per gradient update. Default is 64.}

\item{save.model}{Folder specified to save the model.}
}
\value{
A list with the following components
\item{\code{z}}{Mean of posterior normal distribution, which is
the low-dimensional representation of the cells.}
\item{\code{z.var}}{Log of variance of posterior normal distribution.}
\item{\code{z.samp}}{Random sampling from posterior normal distribution of
the latent characterization.}
\item{\code{B}}{Covariates matrix used to train the model.}
\item{\code{sf}}{Cell specific size factors used.}
\item{\code{adj.raw}}{SAVER-CVAE corrected counts using cdf matching. This
preserves the technical noise characteristics of scRNA-seq expression count
data after adjusting for the observed covariates. This is useful in
downstream scenarios where hypothesis testing on the original count data
is desired, such as in differential expression.}
\item{\code{saver.est}}{Denoised SAVER estimates based on empirical Bayes
method. The SAVER estimate and posterior distribution can be used to
visualize gene expression and perform trajectory analyses. The SAVER
estimate acts as the denoised estimate of the true expression.}
\item{\code{saver.se}}{Standard error of SAVER posterior distribution.}
\item{\code{saver.samp}}{Sampling from the SAVER posterior distribution.
Sampling from the SAVER posterior allows investigation on gene expression
distribution.}
\item{\code{bad.genes}}{Index of genes that could not be predicted accurately
by the autoencoder based on cross validation results. Instead, for these
genes, denoised SAVER estimate are mean gene expression level across cells.}
\item{\code{time}}{Time of decoder training.}
\item{\code{train2}}{A list that contains training information and model
details.}
}
\description{
This is the second step of training SAVER-CVAE.
The decoder takes variational posterior distribution parameters estimated in
the first step as well as observed covariates or library size as input,
and produces the covariate-adjusted parameters of the negative binomial
distribution.
}
\details{
This step starts by sampling from posterior normal distribution of the
latent variable. Then concatenate with observed covariates or scaled library
size as input to the decoder network. The outputs of the decoder are
parameters of the negative binomial distribution.
Cross-validation is used to select the epoch which gives the lowest
validation error.
}
