% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vae_sc_nb.R
\name{run_vae_sc_nb}
\alias{run_vae_sc_nb}
\title{SAVER-CVAE characterizes latent low-dimensional representation using negative
binomial loss}
\usage{
run_vae_sc_nb(
  x,
  B = NULL,
  x.norm = NULL,
  sf = NULL,
  hvg = NULL,
  preprocess = TRUE,
  hvg_genes_only = TRUE,
  n_hvg = 3000,
  perc.exp = 0.01,
  cell_disp = FALSE,
  k = 30,
  alpha = 0.01,
  epochs = NULL,
  batch_size = 64,
  verbose = 2,
  save.model = NULL,
  validation = FALSE,
  val.epochs = 50,
  return_saver = FALSE,
  ...
)
}
\arguments{
\item{x}{Expression count matrix. The rows correspond to genes and colomns
correspond to cells.}

\item{B}{Observed covariates matrix. The rows correspond to cells.
Default is NULL, then scaled log library size will be taken as B.}

\item{x.norm}{A log normailized and standardized gene expression matrix
(cells by genes). If \code{preprocess = TRUE}, \code{x.norm} can be
calculated automatically. Default is NULL.}

\item{sf}{Vector of cell specific size factors. If \code{preprocess = TRUE},
\code{sf} can be calculated automatically. Default is NULL.}

\item{hvg}{Names of highly variable genes used for training the model.
If \code{preprocess = TRUE}, \code{hvg} can be calculated automatically.
Default is NULL.}

\item{preprocess}{Whether to do preprocess before training the model.
TRUE to apply the default \code{preprocess_data} function to data and then
use the output to train the model. Default is TRUE.}

\item{hvg_genes_only}{Whether to train model using only highly variable
genes. TRUE to use only highly variable genes. FALSE to use genes with
non-zero expression in at least perc.exp*100 \% of cells. Default is TRUE.}

\item{n_hvg}{Number of highly variable genes to select and then to train the
model. Default is 3000.}

\item{perc.exp}{Good genes are defined as those with a mean expression of
more than perc.exp. We select highly variable genes from good
genes. If \code{hvg_genes_only = FALSE}, we train model only using good
genes.}

\item{cell_disp}{Whether to use cell-specific dispersion parameter. TRUE is
to use cell-specific dispersion parameter. FALSE is to use batch-specific
dispersion parameter. Default is FALSE.}

\item{k}{Dimension of the latent representation space produced by SAVER-CVAE.
Default is 30.}

\item{alpha}{Weight of reconstruction loss using only observed covariates B.
Default is 0.01.}

\item{epochs}{Number of epochs to train the model in the first step which
aims to characterize the latent variables.
If unspecified, epochs will be 500.}

\item{batch_size}{Number of samples per gradient update. Default is 64.}

\item{verbose}{Verbosity mode. 0 is silent, 1 is progress bar, 2 is one line
per epoch. Default is 2.}

\item{save.model}{Folder specified to save the model.}

\item{validation}{Whether to run the second step of SAVER-CVAE of
training another denoising decoder. Default is FALSE.}

\item{val.epochs}{Number of epochs to train another decoder in the second
step of denoising. Default is 50.}

\item{return_saver}{Whether to calculate corrected expression counts and
SAVER estimates. Default is FALSE.}
}
\value{
A list with the following components
\item{\code{z}}{Mean of posterior normal distribution, which is
the low-dimensional representation of the cells.}
\item{\code{z.var}}{Log of variance of posterior normal distribution.}
\item{\code{z.samp}}{Random generation from posterior normal distribution of
the latent characterization.}
\item{\code{B}}{Covariates matrix used to train the model.}
\item{\code{sf}}{Cell specific size factors used.}
\item{\code{train1}}{A list that contains training information and model
details.}
}
\description{
This is the first step training of SAVER-CVAE using negative binomial loss.
This step aims to find a low-dimensional latent representation of the cells
while adjusting for observed covariates
}
\details{
The SVAER-CVAE model is trained using the standardized log scale gene
expression level and observed covariates (if NULL, just using scaled
library size) as input to the encoder. The prior distribution of latent
variable is standard normal. The learned multivariate normal mean is a
low-dimensional representation of each cell and can be used in visualization
and clustering analyses.
}
