% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaverDenoise.R
\name{SaverDenoise}
\alias{SaverDenoise}
\title{Calculate SAVER-CVAE corrected counts, SAVER estimates and sampled SAVER
estimates.}
\usage{
SaverDenoise(
  SeuratObject,
  assay.name = "RNA",
  slot.name = "counts",
  save.model,
  subset.genes = NULL
)
}
\arguments{
\item{SeuratObject}{A Seurat object.}

\item{assay.name}{Name of assay where count data is stored. Default is 'RNA'.}

\item{slot.name}{Name of slot where raw count data is stored. Default is 'counts'.}

\item{save.model}{Folder where decoder model is stored.}

\item{subset.genes}{Names of a subset of genes to be predicted.
Default is NULL, then all genes will be predicted.}
}
\value{
A Seurat object with three new assays:
\item{\code{saverdenoised}}{Denoised SAVER estimates
based on empirical Bayes
method. The SAVER estimate and posterior distribution can be used to
visualize gene expression and perform trajectory analyses. The SAVER
estimate acts as the denoised estimate of the true expression}
\item{\code{saversamp}}{Sampling from the SAVER posterior
distribution which allows investigation on gene expression
distribution.}
\item{\code{savercovadj}}{SAVER-CVAE corrected counts using
cdf matching. This object
preserves the technical noise characteristics of scRNA-seq expression count
data after adjusting for the observed covariates. It is useful in
downstream scenarios where hypothesis testing on the original count data
is desired, such as in differential expression.}
}
\description{
This function is used for calculating SAVERCAT corrected counts,
SAVER estimates and sampled SAVER estimates.
}
\details{
This function takes a Seurat object and trained decoder model as input, and
returns a Seurat object with three new assays. SAVERCAT corrected counts
are stored in \code{SeuratObject[["savercovadj"]]}. Denoised SAVER estimates
are stored in \code{SeuratObject[["saverdenoised"]]} A sampling from the
SAVER posterior distribution is stored in \code{SeuratObject[["saversamp"]]}.
}
