% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaverPreprocess.R
\name{SaverPreprocess}
\alias{SaverPreprocess}
\title{SAVERCAT setup}
\usage{
SaverPreprocess(
  SeuratObject,
  assay.name = "RNA",
  slot.name = "counts",
  categorical.var = NULL,
  continous.var = NULL,
  n_hvg = 3000,
  hvg_genes_only = TRUE,
  perc.exp = 0.01
)
}
\arguments{
\item{SeuratObject}{A Seurat object.}

\item{assay.name}{Name of assay where count data is stored. Default is 'RNA'.}

\item{slot.name}{Name of slot where raw count data is stored. Default is 'counts'.}

\item{categorical.var}{Names of categorical variables to be adjusted by
SAVERCAT. Default is NULL.}

\item{continous.var}{Names of continuous variables to be adjusted by
SAVERCAT. Default is NULL.}

\item{n_hvg}{Number of highly variable genes to select. Default is 3000.}

\item{hvg_genes_only}{Whether to learn latent representation using only
highly variable genes. TRUE is to use only highly variable genes. FALSE is
to use genes with non-zero expression in at least perc.exp*100 \% of cells.
Default is TRUE.}

\item{perc.exp}{Good genes are defined as those with a mean expression of
more than perc.exp. We select highly variable genes from good genes.}
}
\value{
A Seurat object with information used to train SAVERCAT stored in
\code{@tools$preprocess} as a list with the following components:
\item{\code{x}}{A subset of expression count matrix. The rows correspond to
cells and colomns correspond to highly variable genes or genes with a mean
expression of more than perc.exp.}
\item{\code{x.norm}}{Log normailized and standardized gene expression matrix
(cells by genes).}
\item{\code{sf}}{Cell specific size factors.}
\item{\code{B}}{Observed covariates matrix or scaled log library size.}
\item{\code{hvg}}{Names of highly variable genes.}
}
\description{
This is the preprocessing function used before training SAVERCAT model.
It generates covariates matrix, library size normalization factor,
highly variable genes and other key information required by SAVERCAT.
}
\details{
This function takes a Seurat Object as input. By default, raw counts are
storeed in "counts" \code{slot}, "RNA" \code{assay}. Covariates to be
adjusted are supposed to be a subset of variable in \code{meta.data}.
Names of categorical variables to be adjusted by SAVERCAT are specified by
\code{categorical.var} as a vector.
Names of continuous variable to be adjusted by SAVERCAT are specified by
\code{continous.var} as a vector.
Then covariates matrix is constructed for these variables together with log
library size and sacled.
By default, \code{continous.var=NULL}, \code{categorical.var=NULL}, then only
scaled log library size will be taken as covariate matrix.
This function calculates highly variable genes per batch using
\code{Seurat::FindVariableFeatures} function and takes the union.
Size factor is library size diveded by mean library size. Output of this
function is stored in "tools" \code{slot}.
}
